fix precision problems

+++ src/interfaces/IPendlePriceCapAdapter.sol
@@ -13,14 +13,14 @@ interface IPendlePriceCapAdapter is ICLSynchronicityPriceAdapter {
    * @notice Parameters to create adapter
    * @param assetToUsdAggregator Price feed contract address for (ASSET / USD) pair, can also be CAPO adapter
    * @param pendlePrincipalToken Pendle principal token contract address
-   * @param maxDiscountPerYear Maximum discount per year (in percents), can be up to 1e4 (`PERCENTAGE_FACTOR` = 100%)
+   * @param maxDiscountPerYear Maximum discount per year (in percents), can be up to 1e18 (`PERCENTAGE_FACTOR` = 100%)
    * @param aclManager ACL manager contract
    * @param description Description of the adapter
    */
   struct PendlePriceCapAdapterParams {
     address assetToUsdAggregator;
     address pendlePrincipalToken;
-    uint16 maxDiscountPerYear;
+    uint64 maxDiscountPerYear;
     address aclManager;
     string description;
   }
@@ -30,7 +30,7 @@ interface IPendlePriceCapAdapter is ICLSynchronicityPriceAdapter {
    * @param oldMaxDiscountPerYear Old `_maxDiscountPerYear` value
    * @param newMaxDiscountPerYear New `_maxDiscountPerYear` value
    */
-  event maxDiscountPerYearUpdated(uint16 oldMaxDiscountPerYear, uint16 newMaxDiscountPerYear);
+  event maxDiscountPerYearUpdated(uint64 oldMaxDiscountPerYear, uint64 newMaxDiscountPerYear);
 
   /// @dev Attempted to set zero address
   error ZeroAddress();
@@ -53,7 +53,7 @@ interface IPendlePriceCapAdapter is ICLSynchronicityPriceAdapter {
    * @dev Can be called from risk admin or pool admin only
    * @param maxDiscountPerYear_ New max discount per year
    */
-  function setMaxDiscountPerYear(uint16 maxDiscountPerYear_) external;
+  function setMaxDiscountPerYear(uint64 maxDiscountPerYear_) external;
 
   /**
    * @notice Returns the current discount on PT tokens
